--- src/middlewares/errorHandlers.js
@@ -1,7 +1,0 @@
-'use strict'
-
-function authError(res, msg) {
-    res.status(403).send({ msg: "Authorization error: " + msg });
-}
-
-module.exports = { authError }
