+++ src/services/token.js
@@ -4,6 +4,16 @@ const jwt = require('jsonwebtoken')
 const config = require('../config')
 const {authError} = require('./errorHandlers')
 
+
+function createToken(user) {
+    const payload = { sub: user._id }
+    const token = jwt.sign(payload, config.JWS_SECRET_KEY, {
+        expiresIn: 1440 // One day
+    })
+
+    return token
+}
+
 function verifyToken(req, res, next) {
     const token = req.headers['access-token'] 
 
@@ -20,4 +30,4 @@ function verifyToken(req, res, next) {
     else {authError(res, "There is no token")}
 }
 
-module.exports = {verifyToken}
+module.exports = {createToken, verifyToken}
